Geometry improvements.
The idea of rejecting too long lines sounds good, but I'm not sure about
the threshold. On the other hand, limiting the stretching in too short
lines and adjusting the gap is good.

Also fix up some characters that are simply wasting space. The "default"
font has tabular figures, and the slash has the same width. Now it does
not, and the hyphen is shortened by 25 percent, too.