Fix horizontal offset of gap-less lines.
A gap-less line is centered in the available space. The gap width
is used to calculate the indentation, therefore it must be adjusted
to the remaining space in the line even when no gap is present.

Without this, an overfull line will be shifted to the right, off the
label, by half the default gap width.